Properly shard kv cache in attention
`k_proj` is not properly set sharding hints, so QKVLinear.extend_step cannot
create next `cached_key` with proper hints.
This causes OOM for diffusion model, because the code cannot know the local
batch size.
     Shape: f32[1024,2048,8,128]{3,2,1,0:T(8,128)}
     Unpadded size: 8.00G

To fix it, copy `cached_key.sharding` to `k_proj.sharding`, as `cached_key`
sharding is properly set up.

In addition, this is the reason of RLHF slowdown, so revert the workaround change.
https://github.com/apple/axlearn/pull/849